+++ flutter/test/integrations/mobile_vitals_integration_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_flutter/src/integrations/mobile_vitals_integration.dart';
+import 'package:sentry_flutter/src/sentry_native_state.dart';
 import 'package:sentry_flutter/src/sentry_native_wrapper.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:flutter/scheduler.dart';
@@ -20,7 +21,7 @@ void main() {
 
     test('native app start measurement added to first transaction', () async {
       fixture.options.autoAppStart = false;
-      fixture.options.appStartFinish = DateTime.fromMillisecondsSinceEpoch(10);
+      fixture.state.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
       fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
 
       fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
@@ -39,7 +40,7 @@ void main() {
     test('native app start measurement not added to following transactions',
         () async {
       fixture.options.autoAppStart = false;
-      fixture.options.appStartFinish = DateTime.fromMillisecondsSinceEpoch(10);
+      fixture.state.appStartEnd = DateTime.fromMillisecondsSinceEpoch(10);
       fixture.wrapper.nativeAppStart = NativeAppStart(0, true);
 
       fixture.getMobileVitalsIntegration().call(MockHub(), fixture.options);
@@ -60,9 +61,10 @@ void main() {
 class Fixture {
   final options = SentryFlutterOptions(dsn: fakeDsn);
   final wrapper = MockNativeWrapper();
+  final state = SentryNativeState();
 
   MobileVitalsIntegration getMobileVitalsIntegration() {
-    return MobileVitalsIntegration(wrapper, SchedulerBinding.instance);
+    return MobileVitalsIntegration(wrapper, state, SchedulerBinding.instance);
   }
 
   // ignore: invalid_use_of_internal_member